# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: MemoryOps API
  version: 0.21.0
  description: Control plane API for AI-agent memory ingestion, lifecycle management, retrieval, and audit.
  contact:
    name: MemoryOps
    url: https://github.com/Quazmoz/memoryops
servers:
  - url: http://localhost:8080
security:
  - ApiKeyAuth: []
...
paths:
...
  /v1/ingest/observation:
    post:
      tags: [Observations]
      operationId: ingestObservation
      summary: Ingest an agent-authored observation
      description: >
        Accepts a structured observation from any first-party agent or automation.
        Authenticated by workspace API key (X-API-Key). Idempotent per
        (workspace_id, agent_id, content) using SHA-256. Enqueues to the processor
        pipeline and writes an audit record. Returns 202 Accepted immediately.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ObservationRequest'
            example:
              workspace_id: "018f1234-0000-7000-8000-000000000000"
              content: "Deployment completed in 142 s. All health checks passed."
              agent_id: "deploy-bot"
              user_id: "quinn"
              repo: "org/backend"
              tags: ["deploy", "ci"]
              importance: 0.75
              source_ref: "run/8821"
      responses:
        '202':
          description: Observation accepted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObservationAccepted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimited'
        '413':
          description: Observation body exceeds the configured 1 MiB limit
  /v1/memory:
    get:
      tags: [Memory]
      operationId: listMemory
      summary: List memory units
      description: Returns cursor-paginated memory units filtered within the authenticated workspace.
      parameters:
        - $ref: '#/components/parameters/After'
        - $ref: '#/components/parameters/Limit'
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/MemoryType'
        - name: source
          in: query
          schema:
            $ref: '#/components/schemas/Source'
        - name: tag
          in: query
          schema:
            type: string
        - name: entity
          in: query
          schema:
            type: string
        - name: since
          in: query
          schema:
            type: string
            format: date-time
        - name: as_of
          in: query
          schema:
            type: string
            format: date-time
          description: Reconstruct memory state and decay scores as of this timestamp.
      responses:
        '200':
          description: Paginated memory units
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /v1/memory/search:
    post:
      tags: [Memory]
      operationId: searchMemory
      summary: Search memory units
      description: Runs keyword, vector, or hybrid search with scope, point-in-time, and workspace-pool filters.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemorySearchRequest'
      responses:
        '200':
          description: Ranked memory search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemorySearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /v1/retrieve:
    post:
      tags: [Retrieval]
      operationId: retrieveMemory
      summary: Retrieve token-packed memory context
      description: Runs scoped hybrid retrieval and returns packed memories with a trace.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetrievalRequest'
      responses:
        '200':
          description: Retrieval result with trace
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrievalResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /v1/workspaces:
    post:
      tags: [Workspaces]
      operationId: createWorkspace
      summary: Create a workspace
      description: Creates a tenant workspace that owns memory, keys, integrations, and configuration.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkspaceCreateRequest'
      responses:
        '201':
          description: Workspace created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /v1/workspaces/{id}/stats/history:
    get:
      tags: [Workspaces]
      operationId: getWorkspaceStatsHistory
      summary: Get workspace stats history
      description: Returns daily created, promoted, and soft-deleted memory counts for the authenticated workspace.
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - name: days
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 90
            default: 30
      responses:
        '200':
          description: Workspace activity history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceStatsHistory'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /v1/workspaces/{id}/keys:
    post:
      tags: [API Keys]
      operationId: createApiKey
      summary: Create an API key
      description: Creates a workspace API key and returns the plaintext key once.
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKeyCreateRequest'
      responses:
        '201':
          description: API key created; plaintext key returned once
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyCreateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    get:
      tags: [API Keys]
      operationId: listApiKeys
      summary: List API keys
      description: Lists non-secret API key metadata for a workspace.
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - name: include_revoked
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: API keys
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApiKey'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    After:
      name: after
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    WorkspaceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
    ObservationRequest:
      type: object
      required: [workspace_id, content, agent_id]
      properties:
        workspace_id:
          type: string
          format: uuid
          description: Workspace to ingest into
        content:
          type: string
          minLength: 1
          maxLength: 8000
          description: Observation text
        agent_id:
          type: string
          description: Identifier of the agent authoring this observation
        user_id:
          type: string
          description: Optional user context
        repo:
          type: string
          description: Optional repository context (e.g. "org/repo")
        tags:
          type: array
          maxItems: 20
          items:
            type: string
          description: Optional tags to attach to the resulting memory
        importance:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          description: Override importance score (0.0–1.0). Omit to use rule-table scoring.
        source_ref:
          type: string
          description: Optional external reference (e.g. run id, URL)
    ObservationAccepted:
      type: object
      required: [id, status]
      properties:
        id:
          type: string
          format: uuid
          description: Raw event id
        status:
          type: string
          enum: [queued]
...
    Source:
      type: string
      enum: [github, slack, jira, linear]
    MemoryType:
      type: string
      enum: [episodic, semantic]
    ScopeVisibility:
      type: string
      enum: [private, workspace]
    SearchMode:
      type: string
      enum: [hybrid, vector, keyword]
...
    MemoryScope:
      type: object
      required: [workspace_id]
      properties:
        workspace_id:
          type: string
          format: uuid
        agent_id:
          type: [string, 'null']
        user_id:
          type: [string, 'null']
        repo:
          type: [string, 'null']
...
    MemoryUnit:
      type: object
      required:
        - id
        - workspace_id
        - scope
        - memory_type
        - scope_visibility
        - content
        - entities
        - importance_score
        - importance_overridden
        - source_events
        - decay_score
        - pinned
        - tags
        - version
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
        scope:
          $ref: '#/components/schemas/MemoryScope'
        memory_type:
          $ref: '#/components/schemas/MemoryType'
        scope_visibility:
          $ref: '#/components/schemas/ScopeVisibility'
        content:
          type: string
        entities:
          type: array
          items:
            $ref: '#/components/schemas/Entity'
        importance_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
        importance_overridden:
          type: boolean
        source_events:
          type: array
          items:
            type: string
            format: uuid
        embedding_id:
          type: [string, 'null']
        token_count:
          type: [integer, 'null']
        decay_score:
          type: number
          format: float
        pinned:
          type: boolean
        tags:
          type: array
          items:
            type: string
        version:
          type: integer
        deleted_at:
          type: [string, 'null']
          format: date-time
        last_accessed_at:
          type: [string, 'null']
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    MemoryListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/MemoryUnit'
        next_cursor:
          type: [string, 'null']
    MemorySearchRequest:
      type: object
      required: [query, workspace_id, mode]
      properties:
        query:
          type: string
          minLength: 1
          maxLength: 2000
        workspace_id:
          type: string
          format: uuid
        mode:
          $ref: '#/components/schemas/SearchMode'
        limit:
          type: integer
          minimum: 1
          maximum: 100
        offset:
          type: integer
          minimum: 0
        agent_id:
          type: string
        user_id:
          type: string
        repo:
          type: string
        scope:
          $ref: '#/components/schemas/ScopeFilter'
        memory_types:
          type: array
          items:
            $ref: '#/components/schemas/MemoryType'
        filters:
          $ref: '#/components/schemas/MemorySearchFilters'
        as_of:
          type: string
          format: date-time
        include_workspace_pool:
          type: boolean
          default: false
    ScopeFilter:
      type: object
      properties:
        agent_id:
          type: string
        user_id:
          type: string
        repo:
          type: string
    MemorySearchFilters:
      type: object
      properties:
        memory_type:
          $ref: '#/components/schemas/MemoryType'
        source:
          type: string
        min_importance:
          type: number
          format: float
          minimum: 0
          maximum: 1
        pinned:
          type: boolean
        tags:
          type: array
          items:
            type: string
    MemorySearchResponse:
      type: object
      required: [results, total, query_id]
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/MemorySearchResult'
        total:
          type: integer
          minimum: 0
        query_id:
          type: string
          format: uuid
    MemorySearchResult:
      type: object
      required: [memory, score, rank]
      properties:
        memory:
          $ref: '#/components/schemas/MemoryUnit'
        score:
          type: number
          format: float
        rank:
          type: integer
          minimum: 1
    RetrievalRequest:
      type: object
      required: [query]
      properties:
        scope:
          anyOf:
            - $ref: '#/components/schemas/MemoryScope'
            - type: 'null'
        query:
          type: string
          minLength: 1
          maxLength: 2000
        mode:
          $ref: '#/components/schemas/SearchMode'
        agent_id:
          type: string
        user_id:
          type: string
        repo:
          type: string
        as_of:
          type: string
          format: date-time
        include_workspace_pool:
          type: boolean
          default: false
        token_budget:
          type: integer
          minimum: 256
          maximum: 32768
          default: 4096
        filters:
          anyOf:
            - $ref: '#/components/schemas/RetrievalFilters'
            - type: 'null'
    RetrievalFilters:
      type: object
      properties:
        sources:
          type: array
          items:
            $ref: '#/components/schemas/Source'
        memory_types:
          type: array
          items:
            $ref: '#/components/schemas/MemoryType'
        entities:
          type: array
          items:
            type: string
        since:
          type: string
          format: date-time
        tags:
          type: array
          items:
            type: string
        min_importance:
          type: number
          format: float
    RetrievalResult:
      type: object
      required: [query_id, memories, total_tokens, trace]
      properties:
        query_id:
          type: string
          format: uuid
        memories:
          type: array
          items:
            $ref: '#/components/schemas/ScoredMemory'
        total_tokens:
          type: integer
        trace:
          anyOf:
            - $ref: '#/components/schemas/RetrievalTraceResponse'
            - type: 'null'
    ScoredMemory:
      type: object
      required: [id, content, memory_type, importance_score, decay_score, relevance_score, entities, score_breakdown]
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
        memory_type:
          type: string
        importance_score:
          type: number
          format: float
        decay_score:
          type: number
          format: float
        relevance_score:
          type: number
          format: double
        entities:
          type: array
          items:
            $ref: '#/components/schemas/Entity'
        score_breakdown:
          $ref: '#/components/schemas/ScoreBreakdown'
...
    RetrievalTraceResponse:
      type: object
      required: [query_id, query, as_of, mode, feedback_applied, candidates_evaluated, included_count, excluded_count, entries]
      properties:
        query_id:
          type: string
          format: uuid
        query:
          type: string
        as_of:
          type: [string, 'null']
          format: date-time
        mode:
          $ref: '#/components/schemas/SearchMode'
        feedback_applied:
          type: boolean
        candidates_evaluated:
          type: integer
        included_count:
          type: integer
        excluded_count:
          type: integer
        entries:
          type: array
          items:
            $ref: '#/components/schemas/RetrievalTraceEntry'
    WorkspaceCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
    Workspace:
      type: object
      required: [id, name, config, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        config:
          $ref: '#/components/schemas/WorkspaceConfig'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type: [string, 'null']
          format: date-time
    WorkspaceConfig:
      type: object
      additionalProperties: true
      properties:
        promotion_threshold:
          type: number
          format: float
          minimum: 0.50
          maximum: 1.00
        dedup_cosine_threshold:
          type: number
          format: float
          minimum: 0.80
          maximum: 0.99
        access_count_trigger:
          type: integer
          minimum: 1
        half_life_days:
          type: number
          format: float
          description: Legacy decay configuration retained for backward compatibility.
        decay_rate_episodic:
          type: number
          format: float
        decay_rate_semantic:
          type: number
          format: float
        llm_provider:
          type: [string, 'null']
        llm_model:
          type: [string, 'null']
        embedding_provider:
          type: [string, 'null']
        embedding_model:
          type: [string, 'null']
        decay_half_life_days:
          type: integer
          nullable: true
          minimum: 1
          maximum: 3650
          description: >
            Half-life in days for the memory decay formula.
            Defaults to 30 when null. Lower values cause faster decay.
          example: 30
        pruning_threshold:
          type: number
          format: float
          nullable: true
          minimum: 0.01
          maximum: 0.50
          description: >
            Decay score below which a memory is soft-deleted.
            Defaults to 0.10 when null.
          example: 0.10
        sub_agent_pools:
          type: array
          items:
            type: string
          description: Agent IDs whose published workspace-pool memories are inherited by retrieval/search.
    WorkspaceStatsHistory:
      type: object
      required: [days, series]
      properties:
        days:
          type: integer
          minimum: 1
          maximum: 90
        series:
          type: array
          items:
            $ref: '#/components/schemas/WorkspaceStatsHistoryPoint'
...
    ApiKeyCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
    ApiKeyCreateResponse:
      type: object
      required: [key, api_key]
      properties:
        key:
          type: string
        api_key:
          $ref: '#/components/schemas/ApiKey'
    ApiKey:
      type: object
      required: [id, workspace_id, name, prefix, created_at, revoked]
      properties:
        id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
        name:
          type: string
        prefix:
          type: string
        created_at:
          type: string
          format: date-time
        last_used_at:
          type: [string, 'null']
          format: date-time
        revoked:
          type: boolean
        revoked_at:
          type: [string, 'null']
          format: date-time
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string